:recycle: api: Move error to src
