+++ api/src/certification/enrolment/application/http-error-mapper-configuration.js
@@ -1,12 +1,13 @@
 import { HttpErrors } from '../../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../../shared/application/models/domain-error-mapping-configuration.js';
-import { CertificationCandidateForbiddenDeletionError } from '../domain/errors.js';
+import { CertificationCandidateForbiddenDeletionError, SessionStartedDeletionError } from '../domain/errors.js';
 
 const enrolmentDomainErrorMappingConfiguration = [
   {
     name: CertificationCandidateForbiddenDeletionError.name,
     httpErrorFn: (error) => new HttpErrors.ForbiddenError(error.message, error.code),
   },
+  { name: SessionStartedDeletionError.name, httpErrorFn: (error) => new HttpErrors.ConflictError(error.message) },
 ].map((domainErrorMappingConfiguration) => new DomainErrorMappingConfiguration(domainErrorMappingConfiguration));
 
 export { enrolmentDomainErrorMappingConfiguration };
