+++ api/src/certification/session-management/domain/errors.js
@@ -1,11 +1,5 @@
 import { DomainError } from '../../../shared/domain/errors.js';
 
-class SessionStartedDeletionError extends DomainError {
-  constructor(message = 'La session a déjà commencé.') {
-    super(message);
-  }
-}
-
 class SessionAlreadyFinalizedError extends DomainError {
   constructor(message = 'Cannot finalize session more than once.') {
     super(message);
@@ -55,7 +49,6 @@ export {
   CsvWithNoSessionDataError,
   SessionAlreadyFinalizedError,
   SessionAlreadyPublishedError,
-  SessionStartedDeletionError,
   SessionWithAbortReasonOnCompletedCertificationCourseError,
   SessionWithMissingAbortReasonError,
   SessionWithoutStartedCertificationError,
