+++ api/tests/certification/enrolment/unit/application/http-error-mapper-configuration_test.js
@@ -1,10 +1,13 @@
 import { enrolmentDomainErrorMappingConfiguration } from '../../../../../src/certification/enrolment/application/http-error-mapper-configuration.js';
-import { CertificationCandidateForbiddenDeletionError } from '../../../../../src/certification/enrolment/domain/errors.js';
+import {
+  CertificationCandidateForbiddenDeletionError,
+  SessionStartedDeletionError,
+} from '../../../../../src/certification/enrolment/domain/errors.js';
 import { HttpErrors } from '../../../../../src/shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../../../../src/shared/application/models/domain-error-mapping-configuration.js';
 import { expect } from '../../../../test-helper.js';
 
-describe('Unit | Certification | Session | Application | HttpErrorMapperConfiguration', function () {
+describe('Unit | Certification | Enrolment | Application | HttpErrorMapperConfiguration', function () {
   it('contains a list of HttpErrorMapper instances', function () {
     // given
     // when
@@ -29,6 +32,22 @@ describe('Unit | Certification | Session | Application | HttpErrorMapperConfigur
       expect(error.message).to.equal(
         'Il est interdit de supprimer un candidat de certification déjà lié à un utilisateur.',
       );
+    });
+  });
+
+  context('when mapping "SessionStartedDeletionError"', function () {
+    it('returns an SessionStartedDeletionError Http Error', function () {
+      //given
+      const httpErrorMapper = enrolmentDomainErrorMappingConfiguration.find(
+        (httpErrorMapper) => httpErrorMapper.name === SessionStartedDeletionError.name,
+      );
+
+      //when
+      const error = httpErrorMapper.httpErrorFn(new SessionStartedDeletionError());
+
+      //then
+      expect(error).to.be.instanceOf(HttpErrors.ConflictError);
+      expect(error.message).to.equal('La session a déjà commencé.');
     });
   });
 });
